refactor: split agent_program.rs into module directory
Split the 1,421-line src/agent_program.rs into src/agent_program/ with
logical sub-modules, each ≤ 400 lines:

- types.rs: AgentProgramContext, AgentProgramMemoryRecord, AgentProgram trait
- objective_relay.rs: ObjectiveRelayProgram + tests
- meeting_facilitator.rs: MeetingFacilitatorProgram + StructuredMeetingNotes + tests
- goal_curator.rs: GoalCuratorProgram + StructuredGoalPlan + tests
- improvement_curator.rs: ImprovementCuratorProgram + tests
- parsing.rs: shared goal/status parsing helpers + format functions + tests
- test_support.rs: shared test fixtures (cfg(test) only)
- mod.rs: re-exports all pub items for unchanged external API

All 70 agent_program tests pass. Clippy clean.
